# src/client/swagger.json
{
  "x-generator": "NSwag v13.18.2.0 (NJsonSchema v10.8.0.0 (Newtonsoft.Json v13.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "Shopping APi",
    "description": "Web API for Shopping",
    "contact": {
      "name": "systemad",
      "url": "https://github.com/Systemad/",
      "email": ""
    },
    "version": "v1"
  },
  "host": "localhost:7099",
  "schemes": ["https"],
  "paths": {
    "/cart": {
      "get": {
        "tags": ["ShoppingCart"],
        "operationId": "ShoppingCart_GetShoppingCart",
        "produces": ["text/plain", "application/json", "text/json"],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A list of clients items in shopping cart",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CartItem"
              }
            }
          }
        }
      }
    },
    "/cart/add/{id}/{quantity}": {
      "post": {
        "tags": ["ShoppingCart"],
        "operationId": "ShoppingCart_AddItemToCart",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "integer",
            "name": "quantity",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CartItem"
              }
            }
          }
        }
      }
    },
    "/cart/remove/{id}/{quantity}": {
      "post": {
        "tags": ["ShoppingCart"],
        "operationId": "ShoppingCart_RemoveItemFromCart",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "integer",
            "name": "quantity",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CartItem"
              }
            }
          }
        }
      }
    },
    "/cart/empty": {
      "post": {
        "tags": ["ShoppingCart"],
        "operationId": "ShoppingCart_EmptyCart",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/promotion/all": {
      "get": {
        "tags": ["Promotion"],
        "operationId": "Promotion_GetAllPromotions",
        "produces": ["text/plain", "application/json", "text/json"],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PromotionState"
              }
            }
          }
        }
      }
    },
    "/promotion/all/active": {
      "get": {
        "tags": ["Promotion"],
        "operationId": "Promotion_GetActivePromotions",
        "produces": ["text/plain", "application/json", "text/json"],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PromotionState"
              }
            }
          }
        }
      }
    },
    "/promotion/all/inactive": {
      "get": {
        "tags": ["Promotion"],
        "operationId": "Promotion_GetNonActivePromotions",
        "produces": ["text/plain", "application/json", "text/json"],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PromotionState"
              }
            }
          }
        }
      }
    },
    "/promotion/add": {
      "post": {
        "tags": ["Promotion"],
        "operationId": "Promotion_CreatePromotion",
        "consumes": ["application/json", "text/json", "application/*+json"],
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "name": "promotionCreationDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PromotionCreationDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PromotionState"
              }
            }
          }
        }
      }
    },
    "/promotion/add/{promotionId}/{productId}": {
      "post": {
        "tags": ["Promotion"],
        "operationId": "Promotion_AddProductToPromotion",
        "parameters": [
          {
            "type": "string",
            "name": "promotionId",
            "in": "path",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "productId",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/promotion/remove/{promotionId}/{productId}": {
      "delete": {
        "tags": ["Promotion"],
        "operationId": "Promotion_RemoveProductFromPromotion",
        "parameters": [
          {
            "type": "string",
            "name": "promotionId",
            "in": "path",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "productId",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/promotion/status/{promotionId}/active/{status}": {
      "put": {
        "tags": ["Promotion"],
        "operationId": "Promotion_SetPromotionStatus",
        "parameters": [
          {
            "type": "string",
            "name": "promotionId",
            "in": "path",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "boolean",
            "name": "status",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/product/productId": {
      "get": {
        "tags": ["Product"],
        "summary": "Fetch a product by its ID",
        "operationId": "Product_GetProduct",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "productId",
            "in": "query",
            "description": "The ID of the product",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A ProductDetail object",
            "schema": {
              "$ref": "#/definitions/ProductDetail"
            }
          }
        }
      }
    },
    "/product/productIds": {
      "get": {
        "tags": ["Product"],
        "operationId": "Product_GetProductsById",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "array",
            "name": "productIds",
            "in": "query",
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductDetail"
              }
            }
          }
        }
      }
    },
    "/product/create": {
      "post": {
        "tags": ["Product"],
        "operationId": "Product_CreateProduct",
        "consumes": ["application/json", "text/json", "application/*+json"],
        "parameters": [
          {
            "name": "productCreationDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductCreationDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/product/delete/{productId}": {
      "delete": {
        "tags": ["Product"],
        "operationId": "Product_DeleteProductById",
        "parameters": [
          {
            "type": "string",
            "name": "productId",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/category/all": {
      "get": {
        "tags": ["Category"],
        "summary": "Get all categories",
        "operationId": "Category_GetCategories",
        "produces": ["text/plain", "application/json", "text/json"],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A list of ProductCategory",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/category/{category}": {
      "get": {
        "tags": ["Category"],
        "summary": "Get items of a specific category, default is 10",
        "operationId": "Category_GetItemsForCategory",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "category",
            "in": "path",
            "required": true,
            "description": "Enter a category to fetch items of",
            "x-nullable": true
          },
          {
            "type": "integer",
            "name": "limit",
            "in": "query",
            "description": "Enter amount, leave empty if default amount should be retrieved",
            "format": "int32",
            "default": 10,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "List of products",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductDetail"
              }
            }
          }
        }
      }
    },
    "/category/random": {
      "get": {
        "tags": ["Category"],
        "summary": "Get total items of a random category, default is 10",
        "operationId": "Category_GetItemsOfRandomCategory",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "integer",
            "name": "limit",
            "in": "query",
            "description": "Enter amount, leave empty if default amount should be retrieved",
            "format": "int32",
            "default": 10,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "List of products",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductDetail"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CartItem": {
      "type": "object",
      "required": ["userId", "quantity", "productDetail"],
      "properties": {
        "userId": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "productDetail": {
          "$ref": "#/definitions/ProductDetail"
        }
      }
    },
    "ProductDetail": {
      "type": "object",
      "required": [
        "id",
        "name",
        "description",
        "category",
        "quantity",
        "price",
        "imageUrl",
        "createdAt"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "category": {
          "$ref": "#/definitions/ProductCategory"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "price": {
          "type": "number",
          "format": "decimal"
        },
        "imageUrl": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "ProductCategory": {
      "type": "string",
      "description": "",
      "x-enumNames": [
        "Accessories",
        "Hardware",
        "Software",
        "Books",
        "Movies",
        "Music",
        "Games",
        "Other"
      ],
      "enum": [
        "Accessories",
        "Hardware",
        "Software",
        "Books",
        "Movies",
        "Music",
        "Games",
        "Other"
      ]
    },
    "PromotionState": {
      "type": "object",
      "required": ["id", "name", "productIds", "active"],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "productIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "bannerImg": {
          "type": "string"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "PromotionCreationDto": {
      "type": "object",
      "required": ["name", "active"],
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "bannerImg": {
          "type": "string"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "ProductCreationDto": {
      "type": "object",
      "required": [
        "name",
        "description",
        "category",
        "quantity",
        "price",
        "imageUrl"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "category": {
          "$ref": "#/definitions/ProductCategory"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "price": {
          "type": "number",
          "format": "decimal"
        },
        "imageUrl": {
          "type": "string"
        }
      }
    }
  }
}